refactor: change MoviesPage to movies-page
