+++ src/app/store/reducers/app.reducers.ts
@@ -3,7 +3,7 @@ import { ActionReducerMap } from '@ngrx/store';
 import { routerReducer } from '@ngrx/router-store';
 
 import { AppState } from '../state/app.state';
-import { moviesPageReducers } from './moviesPage.reducers';
+import { moviesPageReducers } from './movies-page.reducers';
 import { showsPageReducers } from './showsPage.reducers';
 import { libraryReducers } from './library.reducers';
